requirements.txt
# Clinical Decision Support Model Requirements
# Python 3.8+

# Core ML/DL frameworks
torch>=2.0.0,<2.2.0
transformers>=4.35.0,<4.40.0
datasets>=2.14.0
accelerate>=0.24.0
sentencepiece>=0.1.99

# Data processing
pandas>=2.0.0
numpy>=1.24.0,<1.26.0
scikit-learn>=1.3.0

# NLP and text processing
nltk>=3.8.0
rouge-score>=0.1.2
sacrebleu>=2.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Model optimization
onnx>=1.14.0
onnxruntime>=1.16.0
# tensorrt  # Install separately based on CUDA version

# Quantization and optimization
optimum>=1.13.0
neural-compressor>=2.3.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.66.0
coloredlogs>=15.0

# Testing and validation
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code quality
ruff>=0.1.0
black>=23.0.0
mypy>=1.5.0
types-PyYAML>=6.0.0
types-requests>=2.31.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Monitoring and logging
tensorboard>=2.14.0
wandb>=0.15.0

# Medical NLP specific (optional)
# scispacy>=0.5.0  # Requires separate installation
# medcat>=1.7.0   # For medical concept extraction

# Development tools
ipykernel>=6.25.0
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.1.0 